+++ rust/sealevel/libraries/test-transaction-utils/src/lib.rs
@@ -1,6 +1,4 @@
-use solana_program::{
-    instruction::Instruction,
-};
+use solana_program::instruction::Instruction;
 use solana_program_test::*;
 use solana_sdk::{
     signature::{Signature, Signer},
