+++ rust/sealevel/libraries/test-utils/Cargo.toml
@@ -7,15 +7,26 @@ edition = "2021"
 no-entrypoint = []
 
 [dependencies]
+borsh = "0.9"
+hyperlane-core = { path = "../../hyperlane-core" }
 hyperlane-sealevel-mailbox = { path = "../../programs/mailbox", features = ["no-entrypoint"] }
-hyperlane-sealevel-test-ism = { path = "../../programs/ism/test-ism", features = ["no-entrypoint"] }
+hyperlane-sealevel-test-ism = { path = "../../programs/ism/test-ism", features = ["test-client"] }
 hyperlane-test-transaction-utils = { path = "../test-transaction-utils" }
+hyperlane-sealevel-message-recipient-interface = { path = "../message-recipient-interface" }
+hyperlane-sealevel-interchain-security-module-interface = { path = "../interchain-security-module-interface" }
+serializable-account-meta = { path = "../serializable-account-meta" }
 solana-program = "=1.14.13"
 solana-program-test = "=1.14.13"
 solana-sdk = "=1.14.13"
 
 [dependencies.spl-token-2022]
 version = "=0.5.0"
+features = ["no-entrypoint"]
+git = "https://github.com/tkporter/eclipse-program-library"
+rev = "2f7e864"
+
+[dependencies.spl-noop]
+version = "=0.1.3"
 features = ["no-entrypoint"]
 git = "https://github.com/tkporter/eclipse-program-library"
 rev = "2f7e864"
