+++ rust/sealevel/programs/hyperlane-sealevel-token/Cargo.toml
@@ -32,6 +32,7 @@ version = "=0.1.3"
 features = ["no-entrypoint"]
 git = "https://github.com/tkporter/eclipse-program-library"
 rev = "2f7e864"
+
 [dependencies.spl-token]
 version = "=3.5.0"
 features = ["no-entrypoint"]
